app/requirements.txt
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
certifi==2023.7.22